+++ Modelos/ModeloReporteVentaEntreFechas.cs
@@ -12,7 +12,7 @@ namespace Modelos
         public string nombre_cliente { get; set; }
         public int cantidad_ventas { get; set; }
         public int cantidad_articulos { get; set; }
-        public float monto_ventas { get; set; }
+        public decimal monto_ventas { get; set; }
 
 
     }
